Settings: Add setting in security for sms message rate alerts from CM10
If we add a preference to control "sms_outgoing_check_max_count" we can
control the message rate by writing to Settings.Secure.SMS_OUTGOING_CHECK_MAX_COUNT.

Currently this is null so it defaults to 30 sms messages in a 15 min period.
Creating and writing this preference allows us to control the sms rate interface.
We can also modify timeframe as well.

This was added to security instead of Mms since this isn't just rate limiting the
sms application.  This rate limits any application that sends texts.

This also creates a new category called Application Security.

Change-Id: I47a2f8d136a4399ebf5477dbd5bb23d83b5693cf

Conflicts:

	res/values/arrays.xml