Settings: fix dinamic tiles NPE
This change ensures that mDynamicUsbTether and mDynamicWifi are not null when
user returns to activity from a subpreference of QuickSettings.

Steps to reproduce the FC:
Go into Settings-> System-> Quick Settings panel-> Tiles and layout, then hit the Back button.
mDynamicUsbTether and mDynamicWifi were removed if they are not supported by the device.

Change-Id: Id871c6d8e422626d2a4ca3243b9ac6bd2a0345e5
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>